pkg/apis/proxy/v1alpha1/package_kojiimagebuild_types.go
/*
Copyright 2020 The OCP Release APIServer Authors.

Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.
*/

package v1alpha1

import (
	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
)

// +genclient
// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object

// +subresource-request
type KojiImageBuildPackage struct {
	metav1.TypeMeta   `json:",inline"`
	metav1.ObjectMeta `json:"metadata,omitempty"`
	Spec              KojiImageBuildPackageSpec   `json:"spec,omitempty"`
	Status            KojiImageBuildPackageStatus `json:"status,omitempty"`
}
type KojiImageBuildPackageSpec struct {
}
type KojiImageBuildPackageStatus struct {
	Packages map[string][]string `json:"packages,omitempty"`
}
